% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_my_graphs.R
\name{list_my_graphs}
\alias{list_my_graphs}
\title{List graphs that are shared with or created by you}
\usage{
list_my_graphs(project_name = NULL)
}
\arguments{
\item{project_name}{filter only to the graphs of the given project, or
specify "all" to view all graphs that are available to you. If not
specified, current active project will be selected}
}
\value{
data.frame with graphs description, uuid, created_by and latest version number
}
\description{
List graphs that are shared with or created by you
}
