% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_graph.R
\name{modify_graph}
\alias{modify_graph}
\title{Modify existing graph, move it into a different project, update description,
create a new version, change dimentions}
\usage{
modify_graph(
  graph_uuid,
  graph = NULL,
  description = NULL,
  dims = NULL,
  project_name = NULL,
  public = NULL
)
}
\arguments{
\item{graph_uuid}{UUID of the graph to modify}

\item{graph}{(optional) Base plot function, ggplot2 or plotly object. If provided will generate a new version of the plot}

\item{description}{(optional) New description of the graph}

\item{dims}{(optional) New width and height in px}

\item{project_name}{(optional) Will move the graph into a project with specified name. The project has to exist!}

\item{public}{Boolean whether the graph should be available via public link}
}
\value{
UUID of the graph
}
\description{
Modify existing graph, move it into a different project, update description,
create a new version, change dimentions
}
