% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_session_info.R
\name{my_session_info}
\alias{my_session_info}
\title{Get current session info like active project and sign-in status.}
\usage{
my_session_info()
}
\value{
Invisibly returns TRUE if session exists
}
\description{
Get current session info like active project and sign-in status.
}
