% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sign_up.R
\name{sign_up}
\alias{sign_up}
\title{Sign up for OmicsChart PREON}
\usage{
sign_up(email, organization_name)
}
\arguments{
\item{email}{The email address of the user.}

\item{organization_name}{The name of the organization or affiliation.}
}
\value{
Invisibly returns `TRUE` if the request was successful.
}
\description{
Sign up for OmicsChart PREON
}
